Extract out fix-all portoin.  Add VB tests
